+++ libsolidity/interface/CompilerStack.cpp
@@ -1045,6 +1045,7 @@ string CompilerStack::createMetadata(Contract const& _contract) const
 		meta["settings"]["optimizer"]["details"] = std::move(details);
 	}
 
+	meta["settings"]["useLiteralContent"] = m_metadataLiteralSources;
 	meta["settings"]["evmVersion"] = m_evmVersion.name();
 	meta["settings"]["compilationTarget"][_contract.contract->sourceUnitName()] =
 		_contract.contract->annotation().canonicalName;